Basic authentication

+++ src/main/java/org/pavlov/service/FileService.java
@@ -2,9 +2,11 @@ package org.pavlov.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.keycloak.KeycloakSecurityContext;
 import org.pavlov.dto.response.FileInfoDto;
 import org.pavlov.dto.response.PageFileResponse;
 import org.pavlov.exception.FileNotFoundException;
+import org.pavlov.exception.ResourceNotFoundException;
 import org.pavlov.mapper.FileMapper;
 import org.pavlov.mapper.FileVersionMapper;
 import org.pavlov.model.File;
@@ -13,6 +15,7 @@ import org.pavlov.model.FileVersion;
 import org.pavlov.model.User;
 import org.pavlov.repository.FileRepository;
 import org.pavlov.repository.FileVersionRepository;
+import org.pavlov.repository.UserRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -26,6 +29,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import static org.pavlov.util.Constant.ERROR_FORBIDDEN;
 import static org.pavlov.util.Constant.ERROR_NOT_FOUND;
 
 @RequiredArgsConstructor
@@ -34,11 +38,13 @@ public class FileService {
 
     private final FileRepository fileRepository;
     private final FileVersionRepository fileVersionRepository;
-    private final UserService userService;
     private final FileMapper fileMapper;
     private final FileVersionMapper fileVersionMapper;
+    private final UserRepository userRepository;
 
-    public PageFileResponse getFileContentByPage(Long fileId, int pageSize, int pageNumber) {
+    public PageFileResponse getFileContentByPage(Long fileId, int pageSize, int pageNumber, String keycloakId) {
+        checkFileAccess(fileId, keycloakId);
+
         byte[] fileContent = findByIdOrThrow(fileId).getData();
 
         String fileContentString = new String(fileContent, StandardCharsets.UTF_8);
@@ -53,20 +59,39 @@ public class FileService {
         return new PageFileResponse(pageContentString, pageNumber, totalPages);
     }
 
-    public File saveFile(MultipartFile file, Optional<LocalDateTime> dateTime) throws IOException {
+    public File saveFile(MultipartFile file, Optional<LocalDateTime> dateTime, String keycloakId) throws IOException {
+
         String fileName = StringUtils.cleanPath(file.getOriginalFilename());
         File fileEntity = new File();
         fileEntity.setName(fileName);
         fileEntity.setData(file.getBytes());
         fileEntity.setType(file.getContentType());
         fileEntity.setDateOfCreation(dateTime.orElse(LocalDateTime.now()));
+
+//        User currentUser = userService.findByKeycloakIdOrThrow(keycloakId);
+        User currentUser = findUserByKeycloakIdOrThrow(keycloakId);
+        fileEntity.setUsers(List.of(currentUser));
+
         return fileRepository.save(fileEntity);
     }
 
+    public void checkFileAccess(Long fileId, String keycloakUserId) {
+        File file = findByIdOrThrow(fileId);
+        boolean hasAccess = file.getUsers().stream()
+                .anyMatch(user -> user.getKeycloakId().equals(keycloakUserId));
+
+        if (!hasAccess) {
+            throw new SecurityException(ERROR_FORBIDDEN );
+        }
+    }
+
     @Transactional
-    public void assignUser(Long id, Long userId) {
+    public void assignUser(Long id, Long userId, String keycloakId) {
+        checkFileAccess(id, keycloakId);
+
         File file = findByIdOrThrow(id);
-        User newUser = userService.findByIdOrThrow(userId);
+//        User newUser = userService.findByIdOrThrow(userId);
+        User newUser = findUserByIdOrThrow(userId);
 
         List<User> userList = file.getUsers();
         userList.add(newUser);
@@ -75,9 +100,12 @@ public class FileService {
     }
 
     @Transactional
-    public void removeUser(Long id, Long userId) {
+    public void removeUser(Long id, Long userId, String keycloakId) {
+        checkFileAccess(id, keycloakId);
+
         File file = findByIdOrThrow(id);
-        User newUser = userService.findByIdOrThrow(userId);
+//        User newUser = userService.findByIdOrThrow(userId);
+        User newUser = findUserByIdOrThrow(userId);
 
         List<User> userList = file.getUsers();
         userList.remove(newUser);
@@ -85,7 +113,8 @@ public class FileService {
         fileRepository.save(file);
     }
 
-    public File getFile(Long id) {
+    public File getFile(Long id, String keycloakId) {
+        checkFileAccess(id, keycloakId);
         return findByIdOrThrow(id);
     }
 
@@ -104,7 +133,8 @@ public class FileService {
     }
 
     @Transactional
-    public void deleteFile(Long id) {
+    public void deleteFile(Long id, String keycloakId) {
+        checkFileAccess(id, keycloakId);
         File file = findByIdOrThrow(id);
         file.setStatus(FileStatus.DELETED);
         fileRepository.save(file);
@@ -112,23 +142,28 @@ public class FileService {
     }
 
     @Transactional
-    public void deleteFileVersion(Long id, long version) {
+    public void deleteFileVersion(Long id, long version, String keycloakId) {
+        checkFileAccess(id, keycloakId);
         fileVersionRepository.deleteByFileIdAndVersion(id, version);
     }
 
     @Transactional
-    public void deleteFileVersions(Long id) {
+    public void deleteFileVersions(Long id, String keycloakId) {
+        checkFileAccess(id, keycloakId);
         fileVersionRepository.deleteByFileId(id);
     }
 
     @Transactional
-    public void deleteAll(Long id) {
+    public void deleteAll(Long id, String keycloakId) {
+        checkFileAccess(id, keycloakId);
         fileVersionRepository.deleteByFileId(id);
         fileRepository.deleteById(id);
     }
 
-    public void updateFileContentOnPage(Long id, int pageNumber, int pageSize, Optional<String> note, String newContent) {
-        File file= findByIdOrThrow(id);
+    public void updateFileContentOnPage(Long id, int pageNumber, int pageSize, Optional<String> note, String newContent, String keycloakId) {
+        checkFileAccess(id, keycloakId);
+
+        File file = findByIdOrThrow(id);
 
         if (note.isPresent()) {
             file.setNote(note.get());
@@ -157,21 +192,22 @@ public class FileService {
         return pages;
     }
 
-    public void updateFileName(Long id, Optional<String> note, String newFileName) {
-        File file = findByIdOrThrow(id);
+    public void updateFileName(Long id, Optional<String> note, String newFileName, String keycloakId) {
+        checkFileAccess(id, keycloakId);
 
+        File file = findByIdOrThrow(id);
         if (note.isPresent()) {
             file.setNote(note.get());
         } else file.setNote("File name updated");
-
         file.setName(newFileName);
         file.setVersion(file.getVersion() + 1);
         fileRepository.save(file);
-
         fileVersionRepository.save(fileVersionMapper.fileToFileVersion(file));
     }
 
-    public void restoreFileVersion(Long id, Long versionId) {
+    public void restoreFileVersion(Long id, Long versionId, String keycloakId) {
+        checkFileAccess(id, keycloakId);
+
         FileVersion fileVersionEntity = fileVersionRepository.findById(versionId)
                 .orElseThrow(() -> new RuntimeException("Version not found"));
 
@@ -183,7 +219,9 @@ public class FileService {
         fileRepository.save(fileEntity);
     }
 
-    public void restoreFile(Long id) {
+    public void restoreFile(Long id, String keycloakId) {
+        checkFileAccess(id, keycloakId);
+
         File file = findByIdOrThrow(id);
         file.setStatus(FileStatus.OK);
         fileRepository.save(file);
@@ -193,5 +231,17 @@ public class FileService {
         return fileRepository.findById(id)
                 .orElseThrow(
                         () -> new FileNotFoundException(ERROR_NOT_FOUND, Long.toString(id)));
+    }
+
+    User findUserByIdOrThrow(Long id) {
+        return userRepository.findById(id)
+                .orElseThrow(
+                        () -> new ResourceNotFoundException("User " + id + " not found"));
+    }
+
+    User findUserByKeycloakIdOrThrow(String id) {
+        return userRepository.findByKeycloakId(id)
+                .orElseThrow(
+                        () -> new ResourceNotFoundException("User with keycloakId " + id + " not found"));
     }
 }
